refactor: 거래내역 서버 Controller 리팩토링
