+++ src/main/java/org/indoles/receiptserviceserver/core/receipt/controller/ReceiptController.java
@@ -1,14 +1,17 @@
 package org.indoles.receiptserviceserver.core.receipt.controller;
 
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
-import org.indoles.receiptserviceserver.core.member.controller.BuyerOnly;
-import org.indoles.receiptserviceserver.core.member.controller.Roles;
-import org.indoles.receiptserviceserver.core.member.controller.SellerOnly;
-import org.indoles.receiptserviceserver.core.member.domain.Role;
-import org.indoles.receiptserviceserver.core.member.dto.SignInInfo;
+
+import org.indoles.receiptserviceserver.core.receipt.controller.interfaces.BuyerOnly;
+import org.indoles.receiptserviceserver.core.receipt.controller.interfaces.Roles;
+import org.indoles.receiptserviceserver.core.receipt.controller.interfaces.SellerOnly;
+import org.indoles.receiptserviceserver.core.receipt.domain.enums.Role;
 import org.indoles.receiptserviceserver.core.receipt.dto.*;
 import org.indoles.receiptserviceserver.core.receipt.service.ReceiptService;
+import org.indoles.receiptserviceserver.global.exception.ErrorCode;
+import org.indoles.receiptserviceserver.global.exception.InfraStructureException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -21,52 +24,81 @@ import java.util.UUID;
 public class ReceiptController {
 
     private final ReceiptService receiptService;
+    private final ObjectMapper objectMapper;
+
 
     /**
      * 거래 내역 조회 API(구매자 전용)
-     * @param buyerInfo
+     *
+     * @param signInInfoString
      * @param offset
      * @param size
      * @return
      */
     @BuyerOnly
     @GetMapping("/buyer")
-    public ResponseEntity<List<BuyerReceiptSimpleInfo>> getReceipts( SignInInfo buyerInfo,
-                                                                    @RequestParam(name = "offset") int offset,
-                                                                    @RequestParam(name = "size") int size) {
+    public ResponseEntity<List<BuyerReceiptSimpleInfo>> getReceipts(
+            @RequestHeader("X-SignIn-Info") String signInInfoString,
+            @RequestParam(name = "offset") int offset,
+            @RequestParam(name = "size") int size
+    ) {
+
+        SignInInfo signInInfo = convertToSignInInfo(signInInfoString);
+
         BuyerReceiptSearchCondition condition = new BuyerReceiptSearchCondition(offset, size);
-        List<BuyerReceiptSimpleInfo> infos = receiptService.getBuyerReceiptSimpleInfos(buyerInfo, condition);
+        List<BuyerReceiptSimpleInfo> infos = receiptService.getBuyerReceiptSimpleInfos(signInInfo, condition);
         return ResponseEntity.ok(infos);
     }
 
     /**
      * 거래 내역 조회 API(판매자 전용)
-     * @param sellerInfo
+     *
+     * @param signInInfoString
      * @param offset
      * @param size
      * @return
      */
     @SellerOnly
     @GetMapping("/seller")
-    public ResponseEntity<List<SellerReceiptSimpleInfo>> getSellerReceipts( SignInInfo sellerInfo,
-                                                                           @RequestParam(name = "offset") int offset,
-                                                                           @RequestParam(name = "size") int size) {
+    public ResponseEntity<List<SellerReceiptSimpleInfo>> getSellerReceipts(
+            @RequestHeader("X-SignIn-Info") String signInInfoString,
+            @RequestParam(name = "offset") int offset,
+            @RequestParam(name = "size") int size
+    ) {
+
+        SignInInfo signInInfo = convertToSignInInfo(signInInfoString);
+
         SellerReceiptSearchCondition condition = new SellerReceiptSearchCondition(offset, size);
-        List<SellerReceiptSimpleInfo> infos = receiptService.getSellerReceiptSimpleInfos(sellerInfo, condition);
+        List<SellerReceiptSimpleInfo> infos = receiptService.getSellerReceiptSimpleInfos(signInInfo, condition);
         return ResponseEntity.ok(infos);
     }
 
     /**
      * 거래 내역 상세 조회 API(구매자, 판매자 공통)
-     * @param memberInfo
+     *
+     * @param signInInfoString
      * @param receiptId
      * @return
      */
     @Roles({Role.BUYER, Role.SELLER})
     @GetMapping("/{receiptId}")
-    public ResponseEntity<ReceiptInfo> getReceipt( SignInInfo memberInfo,
-                                                  @PathVariable("receiptId") UUID receiptId) {
-        ReceiptInfo receiptInfo = receiptService.getReceiptInfo(memberInfo, receiptId);
+    public ResponseEntity<ReceiptInfo> getReceipt(
+            @RequestHeader("X-SignIn-Info") String signInInfoString,
+            @PathVariable("receiptId") UUID receiptId
+    ) {
+
+        SignInInfo signInInfo = convertToSignInInfo(signInInfoString);
+
+
+        ReceiptInfo receiptInfo = receiptService.getReceiptInfo(signInInfo, receiptId);
         return ResponseEntity.ok(receiptInfo);
+    }
+
+    private SignInInfo convertToSignInInfo(String signInInfoString) {
+        try {
+            return objectMapper.readValue(signInInfoString, SignInInfo.class);
+        } catch (Exception e) {
+            throw new InfraStructureException("SignInfo 변환 실패" + e, ErrorCode.A031);
+        }
     }
 }
