+++ src/main/java/org/indoles/receiptserviceserver/global/dto/AuctionRefundRequestMessage.java
@@ -1,7 +1,8 @@
 package org.indoles.receiptserviceserver.global.dto;
 
 
-import org.indoles.receiptserviceserver.core.member.dto.SignInInfo;
+
+import org.indoles.receiptserviceserver.core.receipt.dto.SignInInfo;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
