+++ src/main/java/org/indoles/receiptserviceserver/global/exception/ErrorCode.java
@@ -33,6 +33,7 @@ public enum ErrorCode {
     A028("경매 생성 시, 경매 지속 시간에서 경매 할인 주기 시간을 나누었을때, 나누어 떨어지지 않는 경우 예외가 발생합니다."),
     A029("경매 생성 시, 경매 지속 시간이 정확히 분 단위가 아닌 경우 예외가 발생합니다."),
     A030("경매 입찰 요청 시, 최대 가능 수량을 넘기거나 0이하의 수량을 구매하려고 할때 예외가 발생합니다."),
+    A031("회원 서버로 부터 받은, SignInfo 내용을 객체로 변환하는데 예외가 발생합니다."),
 
     // Receipt 관련 예외 코드
     R000("거래 내역 조회 시, 거래 내역을 찾을 수 없을 경우 예외가 발생합니다."),
