+++ src/main/java/org/indoles/receiptserviceserver/global/util/Mapper.java
@@ -3,12 +3,6 @@ package org.indoles.receiptserviceserver.global.util;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-import org.indoles.receiptserviceserver.core.auction.domain.Auction;
-import org.indoles.receiptserviceserver.core.auction.dto.*;
-import org.indoles.receiptserviceserver.core.auction.entity.AuctionEntity;
-import org.indoles.receiptserviceserver.core.member.domain.Member;
-import org.indoles.receiptserviceserver.core.member.domain.Point;
-import org.indoles.receiptserviceserver.core.member.entity.MemberEntity;
 import org.indoles.receiptserviceserver.core.receipt.domain.Receipt;
 import org.indoles.receiptserviceserver.core.receipt.dto.BuyerReceiptSimpleInfo;
 import org.indoles.receiptserviceserver.core.receipt.dto.ReceiptInfo;
@@ -17,23 +11,6 @@ import org.indoles.receiptserviceserver.core.receipt.entity.ReceiptEntity;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class Mapper {
-
-    public static AuctionInfo convertToAuctionInfo(Auction auction) {
-        return AuctionInfo.builder()
-                .auctionId(auction.getId())
-                .sellerId(auction.getSellerId())
-                .productName(auction.getProductName())
-                .originPrice(auction.getOriginPrice())
-                .currentPrice(auction.getCurrentPrice())
-                .stock(auction.getCurrentStock())
-                .maximumPurchaseLimitCount(auction.getMaximumPurchaseLimitCount())
-                .pricePolicy(auction.getPricePolicy())
-                .variationDuration(auction.getVariationDuration())
-                .startedAt(auction.getStartedAt())
-                .finishedAt(auction.getFinishedAt())
-                .isShowStock(auction.isShowStock())
-                .build();
-    }
 
     public static ReceiptInfo convertToReceiptInfo(Receipt receipt) {
         return ReceiptInfo.builder()
@@ -47,49 +24,6 @@ public final class Mapper {
                 .buyerId(receipt.getBuyerId())
                 .createdAt(receipt.getCreatedAt())
                 .updatedAt(receipt.getUpdatedAt())
-                .build();
-    }
-
-    public static BuyerAuctionInfo convertToBuyerAuctionInfo(Auction auction) {
-        Long currentStock = auction.isShowStock() ? auction.getCurrentStock() : null;
-        Long originStock = auction.isShowStock() ? auction.getOriginStock() : null;
-
-        return BuyerAuctionInfo.builder()
-                .auctionId(auction.getId())
-                .sellerId(auction.getSellerId())
-                .productName(auction.getProductName())
-                .originPrice(auction.getOriginPrice())
-                .currentPrice(auction.getCurrentPrice())
-                .originStock(originStock)
-                .currentStock(currentStock)
-                .maximumPurchaseLimitCount(auction.getMaximumPurchaseLimitCount())
-                .pricePolicy(auction.getPricePolicy())
-                .variationDuration(auction.getVariationDuration())
-                .startedAt(auction.getStartedAt())
-                .finishedAt(auction.getFinishedAt())
-                .build();
-    }
-
-    /**
-     * Auction을 SellerAuctionInfo로 변환 - 현재 판매량 등의 필드를 추가해야 합니다
-     *
-     * @param auction
-     * @return
-     */
-    public static SellerAuctionInfo convertToSellerAuctionInfo(Auction auction) {
-        return SellerAuctionInfo.builder()
-                .auctionId(auction.getId())
-                .productName(auction.getProductName())
-                .originPrice(auction.getOriginPrice())
-                .currentPrice(auction.getCurrentPrice())
-                .originStock(auction.getOriginStock())
-                .currentStock(auction.getCurrentStock())
-                .maximumPurchaseLimitCount(auction.getMaximumPurchaseLimitCount())
-                .pricePolicy(auction.getPricePolicy())
-                .variationDuration(auction.getVariationDuration())
-                .startedAt(auction.getStartedAt())
-                .finishedAt(auction.getFinishedAt())
-                .isShowStock(auction.isShowStock())
                 .build();
     }
 
@@ -115,29 +49,6 @@ public final class Mapper {
                 .build();
     }
 
-    public static BuyerAuctionSimpleInfo convertToBuyerAuctionSimpleInfo(Auction auction) {
-        return new BuyerAuctionSimpleInfo(
-                auction.getId(),
-                auction.getProductName(),
-                auction.getCurrentPrice(),
-                auction.getStartedAt(),
-                auction.getFinishedAt()
-        );
-    }
-
-    public static SellerAuctionSimpleInfo convertToSellerAuctionSimpleInfo(Auction auction) {
-        return new SellerAuctionSimpleInfo(
-                auction.getId(),
-                auction.getProductName(),
-                auction.getOriginPrice(),
-                auction.getCurrentPrice(),
-                auction.getOriginStock(),
-                auction.getCurrentStock(),
-                auction.getStartedAt(),
-                auction.getFinishedAt()
-        );
-    }
-
     public static Receipt convertToReceipt(ReceiptEntity receiptEntity) {
         return Receipt.builder()
                 .id(receiptEntity.getId())
@@ -150,62 +61,6 @@ public final class Mapper {
                 .buyerId(receiptEntity.getBuyerId())
                 .createdAt(receiptEntity.getCreatedAt())
                 .updatedAt(receiptEntity.getUpdatedAt())
-                .build();
-    }
-
-    public static Auction convertToAuction(AuctionEntity auctionEntity) {
-        return new Auction(
-                auctionEntity.getId(),
-                auctionEntity.getSellerId(),
-                auctionEntity.getProductName(),
-                auctionEntity.getOriginPrice(),
-                auctionEntity.getCurrentPrice(),
-                auctionEntity.getOriginStock(),
-                auctionEntity.getCurrentStock(),
-                auctionEntity.getMaximumPurchaseLimitCount(),
-                auctionEntity.getPricePolicy(),
-                auctionEntity.getVariationDuration(),
-                auctionEntity.getStartedAt(),
-                auctionEntity.getFinishedAt(),
-                auctionEntity.isShowStock()
-        );
-    }
-
-    public static AuctionEntity convertToAuctionEntity(Auction auction) {
-        return AuctionEntity.builder()
-                .id(auction.getId())
-                .sellerId(auction.getSellerId())
-                .productName(auction.getProductName())
-                .originPrice(auction.getOriginPrice())
-                .currentPrice(auction.getCurrentPrice())
-                .originStock(auction.getOriginStock())
-                .currentStock(auction.getCurrentStock())
-                .maximumPurchaseLimitCount(auction.getMaximumPurchaseLimitCount())
-                .pricePolicy(auction.getPricePolicy())
-                .variationDuration(auction.getVariationDuration())
-                .startedAt(auction.getStartedAt())
-                .finishedAt(auction.getFinishedAt())
-                .isShowStock(auction.isShowStock())
-                .build();
-    }
-
-    public static Member convertToMember(MemberEntity entity) {
-        return Member.builder()
-                .id(entity.getId())
-                .signInId(entity.getSignInId())
-                .password(entity.getPassword())
-                .role(entity.getRole())
-                .point(new Point(entity.getPoint()))
-                .build();
-    }
-
-    public static MemberEntity convertToMemberEntity(Member member) {
-        return MemberEntity.builder()
-                .id(member.getId())
-                .signInId(member.getSignInId())
-                .password(member.getPassword())
-                .role(member.getRole())
-                .point(member.getPoint().getAmount())
                 .build();
     }
